backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
openai>=1.65.0
aiosqlite>=0.20.0
python-dotenv>=1.0.0
python-multipart>=0.0.20
pypdf>=5.0.0
python-docx>=1.1.0
nh3>=0.2.0
httpx>=0.28.0
